# api/gen/http/openapi.yaml
swagger: "2.0"
info:
    title: Domilike Service
    version: ""
host: localhost:18080
basePath: /api
consumes:
    - application/json
    - application/xml
    - application/gob
produces:
    - application/json
    - application/xml
    - application/gob
paths:
...
    /games/{id}:
        get:
            tags:
                - game
            summary: get game
            operationId: game#get
            parameters:
                - name: id
                  in: path
                  description: Game ID
                  required: true
                  type: integer
            responses:
                "200":
                    description: OK response.
                    schema:
                        type: string
            schemes:
                - http
        delete:
            tags:
                - game
            summary: delete game
            operationId: game#delete
            parameters:
                - name: id
                  in: path
                  description: Game ID
                  required: true
                  type: integer
            responses:
                "204":
                    description: No Content response.
            schemes:
                - http
    /games/{id}/supplies:
        get:
            tags:
                - game
            summary: get_supplies game
            operationId: game#get_supplies
            parameters:
                - name: id
                  in: path
                  description: Game ID
                  required: true
                  type: integer
            responses:
                "200":
                    description: OK response.
                    schema:
                        type: string
            schemes:
                - http
    /games/{id}/trashes:
        get:
            tags:
                - game
            summary: get_trashes game
            operationId: game#get_trashes
            parameters:
                - name: id
                  in: path
                  description: Game ID
                  required: true
                  type: integer
            responses:
                "200":
                    description: OK response.
                    schema:
                        type: string
            schemes:
                - http
...
securityDefinitions:
    jwt_header_Authorization:
        type: apiKey
        description: JWT based authentication
        name: Authorization
        in: header